Upgrade for Java runtimes above Java 8
Because java.xml.bind was removed in Java 9 (released in 2017), this
would not run in runtimes above Java 8.

This adds a compatibility layer. This allows to run this great program
on newer platforms, like ARM macOS.

The target is kept the same, so compatibility with Java 6 should be
kept.
The required JDK was changed to Java 7, since the current Maven does
not support 6 any more.